requirements.txt
Flask==0.11.1
python-jose==1.3.2
requests==2.2.1
